exclusion functionality
