+++ api/roleAccess/exclusionService.ts
@@ -7,15 +7,12 @@ import { http } from '../lib/http';
 import { EnhancedRequest } from '../lib/models';
 import { setHeaders } from '../lib/proxy';
 import { RoleAssignment } from '../user/interfaces/roleAssignment';
-import { CaseRole } from '../workAllocation2/interfaces/caseRole';
 import { RoleCategory } from './models/allocate-role.enum';
 import { CaseRoleRequestPayload, RoleExclusion } from './models/caseRoleRequestPayload';
 import { UserInfo } from 'auth/interfaces/UserInfo';
+import { release2ContentType } from './models/release2ContentType';
 
 const baseRoleAccessUrl = getConfigValue(SERVICES_ROLE_ASSIGNMENT_API_PATH);
-
-export const release2ContentType =
-  'application/vnd.uk.gov.hmcts.role-assignment-service.post-assignment-query-request+json;charset=UTF-8;version=2.0';
 
 export async function findExclusionsForCaseId(req: EnhancedRequest, res: Response, next: NextFunction) {
   const requestPayload = getExclusionRequestPayload(req.body.caseId, req.body.jurisdiction, req.body.caseType);
@@ -135,63 +132,6 @@ export function getExclusionRequestPayload(caseId: string, jurisdiction: string,
       },
     ],
   };
-}
-
-export function getLegalAndJudicialRequestPayload(caseId: string,
-                                                  jurisdiction: string,
-                                                  caseType: string): CaseRoleRequestPayload {
-  return {
-    queryRequests: [
-      {
-          attributes: {
-              caseId: [caseId],
-              caseType: [caseType],
-              jurisdiction: [jurisdiction],
-            },
-          roleCategory: ['LEGAL_OPERATIONS'],
-          roleName: ['case-manager'],
-      },
-    ],
-  };
-}
-
-export async function getRolesByCaseId(req: EnhancedRequest, res: Response, next: NextFunction): Promise<Response> {
-  const requestPayload = getLegalAndJudicialRequestPayload(req.body.caseId, req.body.jurisdiction, req.body.caseType);
-  const basePath = getConfigValue(SERVICES_ROLE_ASSIGNMENT_API_PATH);
-  const fullPath = `${basePath}/am/role-assignments/query`;
-  const headers = setHeaders(req, release2ContentType);
-  try {
-    const response: AxiosResponse = await http.post(fullPath, requestPayload, {headers});
-    const roleExclusions = mapResponseToCaseRoles(response.data.roleAssignmentResponse, req.body.exclusionId, req);
-    return res.status(response.status).send(roleExclusions);
-  } catch (error) {
-    next(error);
-  }
-}
-
-export function mapResponseToCaseRoles(
-  roleAssignments: RoleAssignment[],
-  assignmentId: string,
-  req: EnhancedRequest
-): CaseRole[] {
-  if (assignmentId) {
-    roleAssignments = roleAssignments.filter(roleAssignment => roleAssignment.id === assignmentId);
-  }
-  return roleAssignments.map(roleAssignment => ({
-    actions: [
-      {'id': 'reallocate', 'title': 'Reallocate'},
-      {'id': 'remove', 'title': 'Remove Allocation'},
-    ],
-    actorId: roleAssignment.actorId,
-    email: roleAssignment.actorId ? getEmail(roleAssignment.actorId, req) : null,
-    end: roleAssignment.endTime ? roleAssignment.endTime.toString() : null,
-    id: roleAssignment.id,
-    location: null,
-    name: roleAssignment.actorId ? getUserName(roleAssignment.actorId, req) : null,
-    roleCategory: mapRoleCategory(roleAssignment.roleCategory),
-    roleName: roleAssignment.roleName,
-    start: roleAssignment.beginTime ? roleAssignment.beginTime.toString() : null,
-  }));
 }
 
 export function mapRoleCategory(roleCategory: string): RoleCategory {
