+++ api/roleAccess/models/release2ContentType.ts
@@ -1,0 +1,2 @@
+export const release2ContentType =
+  'application/vnd.uk.gov.hmcts.role-assignment-service.post-assignment-query-request+json;charset=UTF-8;version=2.0';
