+++ api/roleAccess/routes.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 import authInterceptor from '../lib/middleware/auth';
-import { confirmUserExclusion, deleteUserExclusion, findExclusionsForCaseId, getRolesByCaseId } from './exclusionService';
-import { confirmAllocateRole, deleteRoleByCaseAndRoleId, reallocateRole } from './index';
+import { confirmUserExclusion, deleteUserExclusion, findExclusionsForCaseId } from './exclusionService';
+import { confirmAllocateRole, deleteRoleByCaseAndRoleId, getRolesByCaseId, reallocateRole } from './index';
 import { getPossibleRoles } from './roleAssignmentService';
 
 const router = Router({ mergeParams: true });
